Create Cantaloupe deploy using Fargate
We are taking an initial pass at trying this deploy using Fargate and an existing Cantaloupe Docker image. Initially this was just done via the aws cli, and in some cases using existing infrastructure in our AWS account. The basic steps we followed for this are in the README. Specific values that we used have been stripped out, so you'll need to populate these accordingly, for example the <region> and <accountid> values in the deploy/fargatetask.json. These will eventually be driven by inputs or env when we have the CFT/deploy scripts.